requirements.txt
google-cloud-speech==0.29.0
termcolor==1.1.0
emoji==0.4.5
